make crafting use wrapper
