+++ src/main/java/com/cleanroommc/groovyscript/api/INamed.java
@@ -1,0 +1,16 @@
+package com.cleanroommc.groovyscript.api;
+
+import java.util.Collection;
+
+public interface INamed {
+
+    Collection<String> getAliases();
+
+    default String getName() {
+        Collection<String> aliases = getAliases();
+        if (aliases.isEmpty()) {
+            return "EmptyName";
+        }
+        return aliases.iterator().next();
+    }
+}
