+++ src/main/java/com/cleanroommc/groovyscript/compat/vanilla/VanillaModule.java
@@ -26,7 +26,7 @@ public class VanillaModule implements IScriptReloadable {
     public static final InWorldCrafting inWorldCrafting = new InWorldCrafting();
 
     public static void initializeBinding() {
-        GroovyScript.getSandbox().registerBinding("Crafting", crafting);
+        GroovyScript.getSandbox().registerBinding(crafting);
         GroovyScript.getSandbox().registerBinding("Furnace", furnace);
         GroovyScript.getSandbox().registerBinding("Loot", loot);
         GroovyScript.getSandbox().registerBinding("OreDict", oreDict);
@@ -42,6 +42,7 @@ public class VanillaModule implements IScriptReloadable {
     @Override
     @GroovyBlacklist
     public void onReload() {
+        crafting.onReload();
         furnace.onReload();
         loot.onReload();
         oreDict.onReload();
